rpc: Add delay before sending Finish in handleReturn
When handling a Return message, one of the actions required by the
CapNProto protocol is to send a corresponding Finish message, to let the
remote vat know that no more calls will reference the answered question.

However, there is currently a race condition due to lack of proper
synchronization between application code calling a function to a
promised answer (i.e. a question) and the library sending the Finish
message.

In particular, as the TestPromiseOrdering test exposes, due to timing or
system loading, application code that has a reference to the promised
answer may be in the process of queuing a Call message, but that
goroutine may get preempted by the one that is queueing the Finish
message, such that the Finish is sent before the Call.

This causes the remote host to fail the Call, because it would have
already removed its corresponding answer entry (due to Finish arriving
first). This, in turn, causes the aforementioned test to fail.

While this commit does not decisively solve the issue, it adds a stop
gap measure to work around the issue and adds comments explaining it.

The work around is a delay between processing the Return and sending the
Finish, which allows any concurrent Call to be queued before the Finish.
Other calls won't go through the same path, because the Return will
cause the path to be shortened.

Note that this issue does not entirely depend on path shortening and
could happen without it: if a promise is resolved to an import, there's
also the issue with lack of locking that could cause a call to be sent
to the promise (that is being finished) instead of the import.

This will have to be addressed in a more systematic way and this stopgap
work around is being used only in order to improve reliance on the CI by
making the TestPromiseOrdering less flaky.