bbr: Fix test flakes
This fixes test flakes that happen due to timing of clock actions.

The tests involve multiple goroutines that access the current time of
the clock, but it may happen that an Advance() call in the test happens
before a read of the clock time to create a new timer, thus causing a
test flake.

This issue is fixed by adding appropriate synchronization.