rpc: Add workaround in TestPromiseOrdering
Due to lack of proper synchronization, path shortening may cause a local
call to be executed before a remote call.

And the TestPromiseOrdering explicitly asserts on the server that the
order of calls is coming in correctly, thus under load, this test can
flake.

Note that the _reception_ order is correct: the futures that correspond
to the calls end up fulfilled with the correct number, but the _server_
may see the calls out of order.

This may be an issue with embargoes not working correctly or it may be a
misunderstanding of what assumptions can be relied on by application
code using the CapNProto RPC system.

In any case, in order to momentarily improve the reliability of the CI,
we fix the issue by ensuring the path shortening completes before
proceding with future calls.